Bqcore : Small test patch for time series visualization
This was with the functionality of Bisque to show 5d images,
where we have time series microscopic images ( that is usually
 4D x,y,z,c ) . But somehow the time axis on 3d viewer
works only for the first timestamp data (ie. 4D ).

@connor11son helped to debug the endpoint of the data, that has extra
"," , we resolved it in this test patch but it does now solves the
whole problem where in 3D viewer on Bisque it just loads first time
stamp data correct but not the other

We also discussed that maybe it is bug from image service side but need
to look more into it.